pyproject.toml
[project]
name = "policy-explainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "domchao", email = "d.chaopradith@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.59.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "ruff>=0.12.5",
]
